refactor(SentimentSurveyDialog): re-org and port to ts
