+++ client/src/dialogs/SentimentSurvey/SentimentSurveyDialog.scss
